+++ chapter11/src/main/scala/com/reactivedesignpatterns/chapter11/EchoService.scala
@@ -1,4 +1,4 @@
-package com.reactivedesignpatterns.chapter4
+package com.reactivedesignpatterns.chapter11
 
 import akka.actor.Actor
 import akka.actor.ActorRef
