+++ chapter11/src/test/scala/com/reactivedesignpatterns/chapter11/TranslationServiceSpec.scala
@@ -1,4 +1,4 @@
-package com.reactivedesignpatterns.chapter4
+package com.reactivedesignpatterns.chapter11
 
 import org.scalatest.WordSpec
 import org.scalatest.Matchers
